splited codebreaker_web into router and processor
